Add phpstan, add/fix types
The main goal was to provide better param/return types using `class-string` and templates, but I decided to also run PHPStan over the codebase and fix some errors reported. I think this will help to write a better code in the future.

I also had to add symfony/cache as a dev dependency, because the test suite was failing:
```
1) Doctrine2Test: Plain entity
 Test  tests/unit/Codeception/Module/Doctrine2Test.php:testPlainEntity

  [RuntimeException] Setup tool cannot configure caches without doctrine/cache 1.11 or symfony/cache. Please add an explicit dependency to either library.

#1  /Users/anatoly/PhpstormProjects/module-doctrine2/vendor/doctrine/orm/lib/Doctrine/ORM/Tools/Setup.php:184
#2  /Users/anatoly/PhpstormProjects/module-doctrine2/vendor/doctrine/orm/lib/Doctrine/ORM/Tools/Setup.php:160
#3  /Users/anatoly/PhpstormProjects/module-doctrine2/vendor/doctrine/orm/lib/Doctrine/ORM/Tools/Setup.php:139
#4  /Users/anatoly/PhpstormProjects/module-doctrine2/vendor/doctrine/orm/lib/Doctrine/ORM/Tools/Setup.php:87
#5  /Users/anatoly/PhpstormProjects/module-doctrine2/tests/unit/Codeception/Module/Doctrine2Test.php:73
```